Move query back to the request body
